Refactor to source generated
