--- CsvSerializer/Internal/IDeserializer.cs
@@ -1,8 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Csv {
-	internal interface IDeserializer {
-		List<object> Deserialize(IFormatProvider provider, char delimiter, bool skipHeader, ReadOnlyMemory<char> csv);
-	}
-}
