--- CsvSerializer/Internal/INativeConverter.cs
@@ -1,3 +1,0 @@
-﻿namespace Csv.Internal {
-	internal interface INativeConverter<T> : IConverter<T>, IConverterEmitter { }
-}
