--- CsvSerializer/Internal/ISerializer.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using System.Text;
-
-namespace Csv {
-	internal interface ISerializer {
-		void SerializeHeader(char delimiter, StringBuilder stringBuilder);
-		void SerializeItem(IFormatProvider provider, char delimiter, StringBuilder stringBuilder, object item);
-	}
-}
