+++ CsvSerializer/SerializeOrDeserializeSyntaxReceiver.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.CodeAnalysis;
+
+namespace CsvSerializer {
+	internal sealed class SerializeOrDeserializeSyntaxReceiver : ISyntaxContextReceiver {
+		public void OnVisitSyntaxNode(GeneratorSyntaxContext context) {
+
+		}
+	}
+}
