+++ Tests/DynamicSerializerTests.cs
@@ -10,7 +10,7 @@ namespace Tests {
 		public void NullValuesAreSerializedToEmptyColumn() {
 			typeof(ModelWithNullableValues).IsPublic.Should().BeTrue();
 
-			ModelWithNullableValues obj = new ModelWithNullableValues {
+			ModelWithNullableValues obj = new() {
 				Bool = null,
 				Byte = null,
 				SByte = null,
@@ -27,7 +27,10 @@ namespace Tests {
 				DateTime = null
 			};
 			string csv = CsvSerializer.Serialize(new[] { obj }, withHeaders: true);
-			csv.Should().Be("\"Bool\",\"Byte\",\"SByte\",\"Short\",\"UShort\",\"Int\",\"UInt\",\"Long\",\"ULong\",\"Float\",\"Double\",\"Decimal\",\"String\",\"DateTime\"\r\n,,,,,,,,,,,,,");
+			csv.Should().Be("""
+				"Bool","Byte","SByte","Short","UShort","Int","UInt","Long","ULong","Float","Double","Decimal","String","DateTime"
+				,,,,,,,,,,,,,
+				""");
 
 			obj = new ModelWithNullableValues {
 				Bool = true,
@@ -46,7 +49,10 @@ namespace Tests {
 				DateTime = new DateTime(2019, 8, 23)
 			};
 			csv = CsvSerializer.Serialize(new[] { obj }, withHeaders: true);
-			csv.Should().Be("\"Bool\",\"Byte\",\"SByte\",\"Short\",\"UShort\",\"Int\",\"UInt\",\"Long\",\"ULong\",\"Float\",\"Double\",\"Decimal\",\"String\",\"DateTime\"\r\nTrue,102,-100,-200,200,-3000,3000,-40000,40000,1E+14,1.7837193718273812E+19,989898989898,\"CSV Serializer\",\"8/23/2019 12:00:00 AM\"");
+			csv.Should().Be("""
+				"Bool","Byte","SByte","Short","UShort","Int","UInt","Long","ULong","Float","Double","Decimal","String","DateTime"
+				True,102,-100,-200,200,-3000,3000,-40000,40000,1E+14,1.7837193718273812E+19,989898989898,"CSV Serializer","8/23/2019 12:00:00 AM"
+				""");
 		}
 
 		[Fact]
@@ -72,7 +78,10 @@ namespace Tests {
 			item.String.Should().BeEmpty();
 			item.DateTime.Should().BeNull();
 
-			csv = "\"Bool\",\"Byte\",\"SByte\",\"Short\",\"UShort\",\"Int\",\"UInt\",\"Long\",\"ULong\",\"Float\",\"Double\",\"Decimal\",\"String\",\"DateTime\"\r\nTrue,102,-100,-200,200,-3000,3000,-40000,40000,1E+14,1.7837193718273812E+19,989898989898,\"CSV Serializer\",\"08/23/2019 00:00:00\"";
+			csv = """
+				"Bool","Byte","SByte","Short","UShort","Int","UInt","Long","ULong","Float","Double","Decimal","String","DateTime"
+				True,102,-100,-200,200,-3000,3000,-40000,40000,1E+14,1.7837193718273812E+19,989898989898,"CSV Serializer","08/23/2019 00:00:00"
+				""";
 			items = CsvSerializer.Deserialize<ModelWithNullableValues>(csv, hasHeaders: true);
 			items.Length.Should().Be(1);
 			item = items.Single();
@@ -95,17 +104,21 @@ namespace Tests {
 		[Fact]
 		public void DoubleQuotesAreEscapedOnSerializing() {
 			typeof(EscapeTest).IsPublic.Should().BeTrue();
-			EscapeTest obj = new EscapeTest {
+			EscapeTest obj = new() {
 				Name = "Tony \"Iron Man\" Stark"
 			};
 			string csv = CsvSerializer.Serialize(new[] { obj });
-			csv.Should().Be("\"Tony \"\"Iron Man\"\" Stark\"");
+			csv.Should().Be("""
+				"Tony ""Iron Man"" Stark"
+				""");
 		}
 
 		[Fact]
 		public void DoubleQuotesAreUnescapedOnDeserializing() {
 			typeof(EscapeTest).IsPublic.Should().BeTrue();
-			string csv = "\"Tony \"\"Iron Man\"\" Stark\"";
+			string csv = """
+				"Tony ""Iron Man"" Stark"
+				""";
 			EscapeTest[] items = CsvSerializer.Deserialize<EscapeTest>(csv);
 			items.Length.Should().Be(1);
 			EscapeTest item = items[0];
@@ -115,7 +128,10 @@ namespace Tests {
 		[Fact]
 		public void CommasInStringDontSplitString() {
 			typeof(CommaTest).IsPublic.Should().BeTrue();
-			string csv = "\"Stark, Tony\",\"Stark\"\r\n\"Banner, Bruce\",\"Banner\"";
+			string csv = """
+				"Stark, Tony","Stark"
+				"Banner, Bruce","Banner"
+				""";
 			CommaTest[] items = CsvSerializer.Deserialize<CommaTest>(csv);
 			items.Length.Should().Be(2);
 			items[0].Name.Should().Be("Stark, Tony");
